[ALPS04007281] kernel-4.4: Fix P2 build warning
[Detail] fix build warning for callerID

MTK-Commit-Id: e6af8cd0338090eb012759fa43e76e310501e837

CR-Id: ALPS04007281
Feature: [Android Default] Camera Application Basic Functions
(cherry picked from commit 831d47e286b0af3930a6d2ba685b1506698b8f89)

+++ drivers/misc/mediatek/cameraisp/src/mt6739/inc/camera_isp.h
@@ -707,7 +707,7 @@ enum ISP_ED_BUFQUE_CTRL_ENUM {
 struct ISP_ED_BUFQUE_STRUCT {
 	enum ISP_ED_BUFQUE_CTRL_ENUM ctrl;
 	unsigned int processID;
-	unsigned int callerID;
+	unsigned long callerID;
 	int p2burstQIdx;
 	int p2dupCQIdx;
 	int frameNum;
